feat: split descriptor env into separate file
